refactor: extract wallet-trade and workbench route modules from server.ts
- wallet-trade-compat-routes.ts (719 lines): /api/wallet/trade/execute,
  /api/wallet/transfer/execute + trade helpers
- workbench-compat-routes.ts (474 lines): /api/workbench/todos/* CRUD +
  full helper cluster

server.ts: 3,238 → 2,117 lines (-1,121)
Cumulative: 4,269 → 2,117 (-50.4%)
