+++ packages/app-core/src/api/server.ts
@@ -28,7 +28,6 @@ import {
   startApiServer as upstreamStartApiServer,
   validateMcpServerConfig,
 } from "@miladyai/agent/api/server";
-import { type PolicyResult, StewardApiError } from "@stwd/sdk";
 import {
   ensureCompatApiAuthorized,
   ensureCompatSensitiveRouteAuthorized,
@@ -86,20 +85,11 @@ export {
 };
 
 import {
-  buildBscApproveUnsignedTx,
-  buildBscBuyUnsignedTx,
-  buildBscSellUnsignedTx,
-  buildBscTradeQuote,
-  resolveBscApprovalSpender,
-  resolvePrimaryBscRpcUrl,
-} from "@miladyai/agent/api/bsc-trade";
-import {
   getWalletAddresses,
   initStewardWalletCache,
 } from "@miladyai/agent/api/wallet";
 import { fetchEvmNfts } from "@miladyai/agent/api/wallet-evm-balance";
 import { resolveWalletRpcReadiness } from "@miladyai/agent/api/wallet-rpc";
-import { recordWalletTradeLedgerEntry } from "@miladyai/agent/api/wallet-trading-profile";
 import {
   type ElizaConfig,
   loadElizaConfig,
@@ -112,7 +102,6 @@ import {
   sanitizeForSettingsDebug,
   settingsDebugCloudSummary,
 } from "@miladyai/shared";
-import { ethers } from "ethers";
 import {
   ensureRuntimeSqlCompatibility,
   executeRawSql,
@@ -129,6 +118,8 @@ import {
 import { handleAuthPairingCompatRoutes } from "./auth-pairing-compat-routes";
 import { handleDevCompatRoutes } from "./dev-compat-routes";
 import { handlePluginsCompatRoutes } from "./plugins-compat-routes";
+import { handleWalletTradeCompatRoutes } from "./wallet-trade-compat-routes";
+import { handleWorkbenchCompatRoutes } from "./workbench-compat-routes";
 import { resolveDevStackFromEnv } from "./dev-stack";
 import {
   approveStewardTransaction,
@@ -180,18 +171,6 @@ export function isLoopbackRemoteAddress(
     normalized === "::ffff:127.0.0.1" ||
     normalized === "::ffff:0:127.0.0.1"
   );
-}
-
-function resolveWalletExecutionMode(
-  canSign: boolean,
-  canExecuteLocally: boolean,
-  hasStewardSigner: boolean,
-): "local-key" | "steward" | "user-sign" {
-  if (!canSign || !canExecuteLocally) {
-    return "user-sign";
-  }
-
-  return hasStewardSigner ? "steward" : "local-key";
 }
 
 // Lazy-imported to avoid circular dependency with runtime/eliza.ts
@@ -533,23 +512,6 @@ export function hasCompatPersistedOnboardingState(config: ElizaConfig): boolean 
 
 // sendJsonResponse, sendJsonErrorResponse — now imported from ./response
 
-function getStewardPolicyResults(error: StewardApiError): PolicyResult[] {
-  if (
-    error.data &&
-    typeof error.data === "object" &&
-    "results" in error.data &&
-    Array.isArray(error.data.results)
-  ) {
-    return error.data.results as PolicyResult[];
-  }
-
-  return [];
-}
-
-function isStewardPolicyRejection(error: unknown): error is StewardApiError {
-  return error instanceof StewardApiError && error.status === 403;
-}
-
 export function getConfiguredCompatAgentName(): string | null {
   const config = loadElizaConfig();
   const listAgent = config.agents?.list?.[0];
@@ -670,427 +632,6 @@ function patchCompatStatusResponse(
       resolvedCallback,
     );
   }) as typeof res.end;
-}
-
-const WORKBENCH_TODO_TAG = "workbench-todo";
-
-type WorkbenchTodoResponse = {
-  id: string;
-  name: string;
-  description: string;
-  priority: number | null;
-  isUrgent: boolean;
-  isCompleted: boolean;
-  type: string;
-};
-
-function asCompatObject(value: unknown): Record<string, unknown> | null {
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return null;
-  }
-  return value as Record<string, unknown>;
-}
-
-function readCompatTaskMetadata(
-  task: Record<string, unknown>,
-): Record<string, unknown> {
-  return asCompatObject(task.metadata) ?? {};
-}
-
-function normalizeCompatStringArray(value: unknown): string[] {
-  if (!Array.isArray(value)) {
-    return [];
-  }
-
-  return value
-    .filter((entry): entry is string => typeof entry === "string")
-    .map((entry) => entry.trim())
-    .filter((entry) => entry.length > 0);
-}
-
-function parseCompatNullableNumber(value: unknown): number | null {
-  if (value === null || value === undefined || value === "") {
-    return null;
-  }
-
-  if (typeof value === "number" && Number.isFinite(value)) {
-    return Math.trunc(value);
-  }
-
-  if (typeof value === "string") {
-    const parsed = Number.parseInt(value, 10);
-    return Number.isNaN(parsed) ? null : parsed;
-  }
-
-  return null;
-}
-
-function readCompatTaskCompleted(task: Record<string, unknown>): boolean {
-  const metadata = readCompatTaskMetadata(task);
-
-  if (typeof metadata.isCompleted === "boolean") {
-    return metadata.isCompleted;
-  }
-
-  const todoMeta =
-    asCompatObject(metadata.workbenchTodo) ?? asCompatObject(metadata.todo);
-  if (todoMeta && typeof todoMeta.isCompleted === "boolean") {
-    return todoMeta.isCompleted;
-  }
-
-  return false;
-}
-
-function normalizeCompatTodoTags(value: unknown, defaults: string[]): string[] {
-  const tags = new Set(
-    defaults.map((entry) => entry.trim()).filter((entry) => entry.length > 0),
-  );
-
-  for (const tag of normalizeCompatStringArray(value)) {
-    tags.add(tag);
-  }
-
-  return [...tags];
-}
-
-function toTaskBackedWorkbenchTodo(
-  task: Record<string, unknown> | null | undefined,
-): WorkbenchTodoResponse | null {
-  if (!task) {
-    return null;
-  }
-
-  const id =
-    typeof task.id === "string" && task.id.trim().length > 0 ? task.id : null;
-  if (!id) {
-    return null;
-  }
-
-  const tags = new Set(normalizeCompatStringArray(task.tags));
-  const metadata = readCompatTaskMetadata(task);
-  const todoMeta =
-    asCompatObject(metadata.workbenchTodo) ?? asCompatObject(metadata.todo);
-
-  if (!tags.has(WORKBENCH_TODO_TAG) && !tags.has("todo") && !todoMeta) {
-    return null;
-  }
-
-  const name =
-    typeof task.name === "string" && task.name.trim().length > 0
-      ? task.name
-      : "Todo";
-
-  return {
-    id,
-    name,
-    description:
-      typeof todoMeta?.description === "string"
-        ? todoMeta.description
-        : typeof task.description === "string"
-          ? task.description
-          : "",
-    priority: parseCompatNullableNumber(todoMeta?.priority),
-    isUrgent: todoMeta?.isUrgent === true,
-    isCompleted: readCompatTaskCompleted(task),
-    type:
-      typeof todoMeta?.type === "string" && todoMeta.type.trim().length > 0
-        ? todoMeta.type
-        : "task",
-  };
-}
-
-function runtimeHasTodoDatabase(runtime: AgentRuntime | null): boolean {
-  const db = (runtime as { db?: unknown } | null)?.db;
-  return !!db && typeof db === "object";
-}
-
-function decodeCompatTodoId(
-  rawValue: string,
-  res: http.ServerResponse,
-): string | null {
-  try {
-    const decoded = decodeURIComponent(rawValue);
-    if (decoded.trim().length === 0) {
-      sendJsonErrorResponse(res, 400, "Invalid todo id");
-      return null;
-    }
-    return decoded;
-  } catch {
-    sendJsonErrorResponse(res, 400, "Invalid todo id");
-    return null;
-  }
-}
-
-async function handleTaskBackedWorkbenchTodoRoute(
-  req: http.IncomingMessage,
-  res: http.ServerResponse,
-  runtime: AgentRuntime | null,
-  pathname: string,
-  method: string,
-): Promise<boolean> {
-  if (!runtime) {
-    return false;
-  }
-
-  if (
-    pathname !== "/api/workbench/todos" &&
-    !pathname.startsWith("/api/workbench/todos/")
-  ) {
-    return false;
-  }
-
-  if (!ensureCompatApiAuthorized(req, res)) {
-    return true;
-  }
-
-  let operation = "route";
-  try {
-    const getTaskList = async () =>
-      (
-        (await runtime.getTasks({})) as unknown as Array<
-          Record<string, unknown>
-        >
-      ).map((task) => task as Record<string, unknown>);
-
-    if (method === "GET" && pathname === "/api/workbench/todos") {
-      operation = "list todos";
-      const todos = (await getTaskList())
-        .map((task) => toTaskBackedWorkbenchTodo(task))
-        .filter((todo): todo is WorkbenchTodoResponse => todo !== null)
-        .sort((left, right) => left.name.localeCompare(right.name));
-
-      sendJsonResponse(res, 200, { todos });
-      return true;
-    }
-
-    if (method === "POST" && pathname === "/api/workbench/todos") {
-      const body = await readCompatJsonBody(req, res);
-      if (body == null) {
-        return true;
-      }
-
-      const name = typeof body.name === "string" ? body.name.trim() : "";
-      if (!name) {
-        sendJsonErrorResponse(res, 400, "name is required");
-        return true;
-      }
-
-      const description =
-        typeof body.description === "string" ? body.description : "";
-      const type =
-        typeof body.type === "string" && body.type.trim().length > 0
-          ? body.type.trim()
-          : "task";
-
-      operation = "create todo";
-      const taskId = await runtime.createTask({
-        name,
-        description,
-        tags: normalizeCompatTodoTags(body.tags, [WORKBENCH_TODO_TAG, "todo"]),
-        metadata: {
-          isCompleted: false,
-          workbenchTodo: {
-            description,
-            priority: parseCompatNullableNumber(body.priority),
-            isUrgent: body.isUrgent === true,
-            isCompleted: false,
-            type,
-          },
-        },
-      });
-
-      operation = "load created todo";
-      const created = await runtime.getTask(taskId);
-      const todo = toTaskBackedWorkbenchTodo(
-        created as Record<string, unknown> | null,
-      );
-      if (!todo) {
-        sendJsonErrorResponse(res, 500, "Todo created but unavailable");
-        return true;
-      }
-
-      sendJsonResponse(res, 201, { todo });
-      return true;
-    }
-
-    const todoCompleteMatch =
-      /^\/api\/workbench\/todos\/([^/]+)\/complete$/.exec(pathname);
-    if (method === "POST" && todoCompleteMatch) {
-      const todoId = decodeCompatTodoId(todoCompleteMatch[1], res);
-      if (!todoId) {
-        return true;
-      }
-
-      const body = await readCompatJsonBody(req, res);
-      if (body == null) {
-        return true;
-      }
-
-      operation = "load todo for completion";
-      const todoTask = (await runtime.getTask(todoId)) as Record<
-        string,
-        unknown
-      > | null;
-      const todo = toTaskBackedWorkbenchTodo(todoTask);
-      if (!todoTask || !todo) {
-        sendJsonErrorResponse(res, 404, "Todo not found");
-        return true;
-      }
-
-      const metadata = readCompatTaskMetadata(todoTask);
-      const todoMeta =
-        asCompatObject(metadata.workbenchTodo) ?? asCompatObject(metadata.todo);
-      const isCompleted = body.isCompleted === true;
-
-      operation = "update todo completion";
-      await runtime.updateTask(todoId, {
-        metadata: {
-          ...metadata,
-          isCompleted,
-          workbenchTodo: {
-            ...(todoMeta ?? {}),
-            isCompleted,
-          },
-        },
-      });
-
-      sendJsonResponse(res, 200, { ok: true });
-      return true;
-    }
-
-    const todoItemMatch = /^\/api\/workbench\/todos\/([^/]+)$/.exec(pathname);
-    if (!todoItemMatch) {
-      return false;
-    }
-
-    const todoId = decodeCompatTodoId(todoItemMatch[1], res);
-    if (!todoId) {
-      return true;
-    }
-
-    if (method === "GET") {
-      operation = "load todo";
-      const todoTask = (await runtime.getTask(todoId)) as Record<
-        string,
-        unknown
-      > | null;
-      const todo = toTaskBackedWorkbenchTodo(todoTask);
-      if (!todoTask || !todo) {
-        sendJsonErrorResponse(res, 404, "Todo not found");
-        return true;
-      }
-
-      sendJsonResponse(res, 200, { todo });
-      return true;
-    }
-
-    if (method === "DELETE") {
-      operation = "load todo for deletion";
-      const todoTask = (await runtime.getTask(todoId)) as Record<
-        string,
-        unknown
-      > | null;
-      if (!todoTask || !toTaskBackedWorkbenchTodo(todoTask)) {
-        sendJsonErrorResponse(res, 404, "Todo not found");
-        return true;
-      }
-
-      operation = "delete todo";
-      await runtime.deleteTask(todoId);
-      sendJsonResponse(res, 200, { ok: true });
-      return true;
-    }
-
-    if (method === "PUT") {
-      const body = await readCompatJsonBody(req, res);
-      if (body == null) {
-        return true;
-      }
-
-      operation = "load todo for update";
-      const todoTask = (await runtime.getTask(todoId)) as Record<
-        string,
-        unknown
-      > | null;
-      const existingTodo = toTaskBackedWorkbenchTodo(todoTask);
-      if (!todoTask || !existingTodo) {
-        sendJsonErrorResponse(res, 404, "Todo not found");
-        return true;
-      }
-
-      if (typeof body.name === "string" && body.name.trim().length === 0) {
-        sendJsonErrorResponse(res, 400, "name cannot be empty");
-        return true;
-      }
-
-      const metadata = readCompatTaskMetadata(todoTask);
-      const todoMeta =
-        asCompatObject(metadata.workbenchTodo) ?? asCompatObject(metadata.todo);
-      const nextTodoMeta: Record<string, unknown> = {
-        ...(todoMeta ?? {}),
-      };
-      const update: Record<string, unknown> = {};
-
-      if (typeof body.name === "string") {
-        update.name = body.name.trim();
-      }
-      if (typeof body.description === "string") {
-        update.description = body.description;
-        nextTodoMeta.description = body.description;
-      }
-      if (body.priority !== undefined) {
-        nextTodoMeta.priority = parseCompatNullableNumber(body.priority);
-      }
-      if (typeof body.isUrgent === "boolean") {
-        nextTodoMeta.isUrgent = body.isUrgent;
-      }
-      if (typeof body.type === "string" && body.type.trim().length > 0) {
-        nextTodoMeta.type = body.type.trim();
-      }
-      if (body.tags !== undefined) {
-        update.tags = normalizeCompatTodoTags(body.tags, [
-          WORKBENCH_TODO_TAG,
-          "todo",
-        ]);
-      }
-
-      const isCompleted =
-        typeof body.isCompleted === "boolean"
-          ? body.isCompleted
-          : existingTodo.isCompleted;
-      nextTodoMeta.isCompleted = isCompleted;
-
-      update.metadata = {
-        ...metadata,
-        isCompleted,
-        workbenchTodo: nextTodoMeta,
-      };
-
-      operation = "update todo";
-      await runtime.updateTask(todoId, update);
-
-      operation = "load updated todo";
-      const refreshed = await runtime.getTask(todoId);
-      const todo = toTaskBackedWorkbenchTodo(
-        refreshed as Record<string, unknown> | null,
-      );
-      if (!todo) {
-        sendJsonErrorResponse(res, 500, "Todo updated but unavailable");
-        return true;
-      }
-
-      sendJsonResponse(res, 200, { todo });
-      return true;
-    }
-
-    return false;
-  } catch (err) {
-    logger.error(
-      `[workbench/todos] ${operation} failed: ${err instanceof Error ? err.message : String(err)}`,
-    );
-    sendJsonErrorResponse(res, 500, `Failed to ${operation}`);
-    return true;
-  }
 }
 
 async function _getTableColumnNames(
@@ -1305,10 +846,6 @@ async function handleDatabaseRowsCompatRoute(
   return true;
 }
 
-type TradePermissionMode = "user-sign-only" | "manual-local-key" | "agent-auto";
-
-const AGENT_AUTOMATION_HEADER = "x-milady-agent-action";
-
 /**
  * Build the set of localhost ports allowed for CORS.
  * Reads from env vars at call time so tests can override.
@@ -1341,97 +878,6 @@ export function isAllowedLocalOrigin(
   } catch {
     return false;
   }
-}
-
-export function resolveTradePermissionMode(config: {
-  features?: { tradePermissionMode?: unknown } | null;
-}): TradePermissionMode {
-  const raw = config.features?.tradePermissionMode;
-  if (
-    raw === "user-sign-only" ||
-    raw === "manual-local-key" ||
-    raw === "agent-auto"
-  ) {
-    return raw;
-  }
-  return "user-sign-only";
-}
-
-export function canUseLocalTradeExecution(
-  mode: TradePermissionMode,
-  isAgent: boolean,
-): boolean {
-  if (mode === "agent-auto") {
-    return true;
-  }
-  if (mode === "manual-local-key") {
-    return !isAgent;
-  }
-  return false;
-}
-
-function isAgentAutomationRequest(
-  req: Pick<http.IncomingMessage, "headers">,
-): boolean {
-  const raw = req.headers[AGENT_AUTOMATION_HEADER];
-  return typeof raw === "string" && /^(1|true|yes|agent)$/i.test(raw.trim());
-}
-
-interface LocalSignedTransactionResult {
-  hash: string;
-  nonce: number;
-  gasLimit: string;
-}
-
-/** @deprecated Use signTransactionWithOptionalSteward() via steward-bridge instead. */
-async function _sendLocalWalletTransaction(
-  rpcUrl: string,
-  tx: {
-    to: string;
-    data?: string;
-    value: bigint;
-    chainId: number;
-    nonce?: number;
-  },
-): Promise<LocalSignedTransactionResult> {
-  const evmKey = process.env.EVM_PRIVATE_KEY ?? "";
-  const provider = new ethers.JsonRpcProvider(rpcUrl);
-
-  try {
-    const wallet = new ethers.Wallet(
-      evmKey.startsWith("0x") ? evmKey : `0x${evmKey}`,
-      provider,
-    );
-    const txResponse = await wallet.sendTransaction(tx);
-    return {
-      hash: txResponse.hash,
-      nonce: txResponse.nonce,
-      gasLimit: txResponse.gasLimit?.toString() ?? "0",
-    };
-  } finally {
-    provider.destroy();
-  }
-}
-
-function resolveBscExecutionNetwork(): {
-  chainId: number;
-  explorerBaseUrl: string;
-} {
-  if (process.env.MILADY_WALLET_NETWORK?.trim().toLowerCase() === "testnet") {
-    const parsedChainId = Number.parseInt(
-      process.env.BSC_TESTNET_CHAIN_ID?.trim() ?? "97",
-      10,
-    );
-    return {
-      chainId: Number.isNaN(parsedChainId) ? 97 : parsedChainId,
-      explorerBaseUrl: "https://testnet.bscscan.com",
-    };
-  }
-
-  return {
-    chainId: 56,
-    explorerBaseUrl: "https://bscscan.com",
-  };
 }
 
 /**
@@ -1551,19 +997,8 @@ async function handleMiladyCompatRoute(
     return false;
   }
 
-  if (
-    !runtimeHasTodoDatabase(state.current) &&
-    url.pathname.startsWith("/api/workbench/todos") &&
-    (await handleTaskBackedWorkbenchTodoRoute(
-      req,
-      res,
-      state.current,
-      url.pathname,
-      method,
-    ))
-  ) {
-    return true;
-  }
+  // Workbench / todos routes — extracted to workbench-compat-routes.ts
+  if (await handleWorkbenchCompatRoutes(req, res, state)) return true;
 
   // Handle all /api/cloud/* routes (except compat and billing which have
   // their own handlers above) through Milady's handleCloudRoute. This is
@@ -2298,564 +1733,8 @@ async function handleMiladyCompatRoute(
     return true;
   }
 
-  if (method === "POST" && url.pathname === "/api/wallet/trade/execute") {
-    if (!ensureCompatApiAuthorized(req, res)) {
-      return true;
-    }
-
-    const body = await readCompatJsonBody(req, res);
-    if (body == null) {
-      return true;
-    }
-
-    const side = typeof body.side === "string" ? body.side : "";
-    const tokenAddress =
-      typeof body.tokenAddress === "string" ? body.tokenAddress : "";
-    const amount = typeof body.amount === "string" ? body.amount : "";
-    const routeProvider =
-      body.routeProvider === "0x" ||
-      body.routeProvider === "pancakeswap-v2" ||
-      body.routeProvider === "auto"
-        ? body.routeProvider
-        : undefined;
-
-    if (!side || !tokenAddress || !amount) {
-      sendJsonErrorResponse(
-        res,
-        400,
-        "side, tokenAddress, and amount are required",
-      );
-      return true;
-    }
-
-    if (side !== "buy" && side !== "sell") {
-      sendJsonErrorResponse(res, 400, 'side must be "buy" or "sell"');
-      return true;
-    }
-
-    const config = loadElizaConfig();
-    const tradePermissionMode = _resolveTradePermissionMode(config);
-    const canExecuteLocally = _canUseLocalTradeExecution(
-      tradePermissionMode,
-      isAgentAutomationRequest(req),
-    );
-    const addresses = getWalletAddresses();
-    const walletAddress = addresses.evmAddress ?? null;
-    const hasLocalKey = Boolean(process.env.EVM_PRIVATE_KEY?.trim());
-    const hasStewardSigner = isStewardConfigured();
-    const canSign = hasLocalKey || hasStewardSigner;
-    const rpcReadiness = resolveWalletRpcReadiness(config);
-    const bscExecutionNetwork = resolveBscExecutionNetwork();
-
-    try {
-      const quote = await buildBscTradeQuote({
-        walletAddress,
-        rpcUrls: rpcReadiness.bscRpcUrls,
-        cloudManagedAccess: rpcReadiness.cloudManagedAccess,
-        request: {
-          side,
-          tokenAddress,
-          amount,
-          slippageBps:
-            typeof body.slippageBps === "number" ? body.slippageBps : undefined,
-          routeProvider,
-        },
-      });
-
-      const unsignedTx =
-        quote.side === "buy"
-          ? buildBscBuyUnsignedTx(
-              quote,
-              walletAddress,
-              typeof body.deadlineSeconds === "number"
-                ? body.deadlineSeconds
-                : undefined,
-            )
-          : buildBscSellUnsignedTx(
-              quote,
-              walletAddress,
-              typeof body.deadlineSeconds === "number"
-                ? body.deadlineSeconds
-                : undefined,
-            );
-
-      let unsignedApprovalTx:
-        | ReturnType<typeof buildBscApproveUnsignedTx>
-        | undefined;
-      let requiresApproval = false;
-      if (quote.side === "sell" && walletAddress) {
-        unsignedApprovalTx = buildBscApproveUnsignedTx(
-          quote.tokenAddress,
-          walletAddress,
-          resolveBscApprovalSpender(quote),
-          quote.quoteIn.amountWei,
-        );
-        requiresApproval = true;
-      }
-
-      if (!canSign || !canExecuteLocally || body.confirm !== true) {
-        sendJsonResponse(res, 200, {
-          ok: true,
-          side: quote.side,
-          mode: resolveWalletExecutionMode(
-            canSign,
-            canExecuteLocally,
-            hasStewardSigner,
-          ),
-          quote,
-          executed: false,
-          requiresUserSignature: true,
-          unsignedTx,
-          unsignedApprovalTx,
-          requiresApproval,
-        });
-        return true;
-      }
-
-      const rpcUrl = resolvePrimaryBscRpcUrl({
-        rpcUrls: rpcReadiness.bscRpcUrls,
-        cloudManagedAccess: rpcReadiness.cloudManagedAccess,
-      });
-
-      let approvalHash: string | undefined;
-      let finalHash = "";
-      let finalNonce: number | null = null;
-      let finalGasLimit = "0";
-      let finalMode: "local-key" | "steward" = hasLocalKey
-        ? "local-key"
-        : "steward";
-
-      if (hasLocalKey && canExecuteLocally) {
-        if (!rpcUrl) {
-          sendJsonErrorResponse(
-            res,
-            503,
-            "BSC RPC not configured for local execution",
-          );
-          return true;
-        }
-
-        if (requiresApproval && unsignedApprovalTx) {
-          const approvalResult = await _sendLocalWalletTransaction(rpcUrl, {
-            to: unsignedApprovalTx.to,
-            data: unsignedApprovalTx.data,
-            value: BigInt(unsignedApprovalTx.valueWei),
-            chainId: unsignedApprovalTx.chainId,
-          });
-          approvalHash = approvalResult.hash;
-          const provider = new ethers.JsonRpcProvider(rpcUrl);
-          try {
-            await provider.waitForTransaction(approvalHash, 1);
-          } finally {
-            provider.destroy();
-          }
-        }
-
-        const localExecution = await _sendLocalWalletTransaction(rpcUrl, {
-          to: unsignedTx.to,
-          data: unsignedTx.data,
-          value: BigInt(unsignedTx.valueWei),
-          chainId: unsignedTx.chainId,
-        });
-        finalHash = localExecution.hash;
-        finalNonce = localExecution.nonce;
-        finalGasLimit = localExecution.gasLimit;
-      } else {
-        finalMode = "steward";
-        if (requiresApproval && unsignedApprovalTx) {
-          const approvalResult = await signTransactionWithOptionalSteward({
-            evmAddress: walletAddress,
-            tx: {
-              to: unsignedApprovalTx.to,
-              data: unsignedApprovalTx.data,
-              value: unsignedApprovalTx.valueWei,
-              chainId: unsignedApprovalTx.chainId,
-              broadcast: true,
-            },
-          });
-
-          if (
-            approvalResult.mode === "steward" &&
-            approvalResult.pendingApproval
-          ) {
-            sendJsonResponse(res, 200, {
-              ok: true,
-              side: quote.side,
-              mode: "steward",
-              quote,
-              executed: false,
-              requiresUserSignature: false,
-              unsignedTx,
-              unsignedApprovalTx,
-              requiresApproval,
-              approval: {
-                status: "pending_approval",
-                policyResults: approvalResult.policyResults,
-              },
-            });
-            return true;
-          }
-
-          approvalHash =
-            "txHash" in approvalResult ? approvalResult.txHash : "";
-
-          if (approvalResult.mode === "steward" && rpcUrl) {
-            const provider = new ethers.JsonRpcProvider(rpcUrl);
-            try {
-              await provider.waitForTransaction(approvalHash, 1);
-            } finally {
-              provider.destroy();
-            }
-          }
-        }
-
-        const executionResult = await signTransactionWithOptionalSteward({
-          evmAddress: walletAddress,
-          tx: {
-            to: unsignedTx.to,
-            data: unsignedTx.data,
-            value: unsignedTx.valueWei,
-            chainId: unsignedTx.chainId,
-            broadcast: true,
-          },
-        });
-
-        if (
-          executionResult.mode === "steward" &&
-          executionResult.pendingApproval
-        ) {
-          sendJsonResponse(res, 200, {
-            ok: true,
-            side: quote.side,
-            mode: "steward",
-            quote,
-            executed: false,
-            requiresUserSignature: false,
-            unsignedTx,
-            unsignedApprovalTx,
-            requiresApproval,
-            approvalHash,
-            execution: {
-              status: "pending_approval",
-              policyResults: executionResult.policyResults,
-            },
-          });
-          return true;
-        }
-
-        finalHash = "txHash" in executionResult ? executionResult.txHash : "";
-      }
-
-      try {
-        const tradeSource =
-          body.source === "agent" || body.source === "manual"
-            ? body.source
-            : "manual";
-
-        recordWalletTradeLedgerEntry({
-          hash: finalHash,
-          source: tradeSource,
-          side: quote.side,
-          tokenAddress: quote.tokenAddress,
-          slippageBps: quote.slippageBps,
-          route: quote.route,
-          quoteIn: {
-            symbol: quote.quoteIn.symbol,
-            amount: quote.quoteIn.amount,
-            amountWei: quote.quoteIn.amountWei,
-          },
-          quoteOut: {
-            symbol: quote.quoteOut.symbol,
-            amount: quote.quoteOut.amount,
-            amountWei: quote.quoteOut.amountWei,
-          },
-          status: "pending",
-          confirmations: 0,
-          nonce: finalNonce,
-          blockNumber: null,
-          gasUsed: null,
-          effectiveGasPriceWei: null,
-          explorerUrl: `${bscExecutionNetwork.explorerBaseUrl}/tx/${finalHash}`,
-        });
-      } catch (ledgerErr) {
-        logger.warn(
-          `[api] Failed to record trade ledger entry: ${ledgerErr instanceof Error ? ledgerErr.message : ledgerErr}`,
-        );
-      }
-
-      sendJsonResponse(res, 200, {
-        ok: true,
-        side: quote.side,
-        mode: finalMode,
-        quote,
-        executed: true,
-        requiresUserSignature: false,
-        unsignedTx,
-        unsignedApprovalTx,
-        requiresApproval,
-        execution: {
-          hash: finalHash,
-          nonce: finalNonce,
-          gasLimit: finalGasLimit,
-          valueWei: unsignedTx.valueWei,
-          explorerUrl: `${bscExecutionNetwork.explorerBaseUrl}/tx/${finalHash}`,
-          blockNumber: null,
-          status: "pending",
-          approvalHash,
-        },
-      });
-    } catch (err) {
-      if (isStewardPolicyRejection(err)) {
-        sendJsonResponse(res, 403, {
-          ok: false,
-          mode: "steward",
-          executed: false,
-          requiresUserSignature: false,
-          error: err.message,
-          execution: {
-            status: "rejected",
-            policyResults: getStewardPolicyResults(err),
-          },
-        });
-        return true;
-      }
-
-      sendJsonErrorResponse(
-        res,
-        500,
-        `Trade execution failed: ${err instanceof Error ? err.message : "unknown error"}`,
-      );
-    }
-    return true;
-  }
-
-  if (method === "POST" && url.pathname === "/api/wallet/transfer/execute") {
-    if (!ensureCompatApiAuthorized(req, res)) {
-      return true;
-    }
-
-    const body = await readCompatJsonBody(req, res);
-    if (body == null) {
-      return true;
-    }
-
-    const toAddressRaw =
-      typeof body.toAddress === "string" ? body.toAddress.trim() : "";
-    const amount = typeof body.amount === "string" ? body.amount.trim() : "";
-    const assetSymbol =
-      typeof body.assetSymbol === "string" ? body.assetSymbol.trim() : "";
-
-    if (!toAddressRaw || !amount || !assetSymbol) {
-      sendJsonErrorResponse(
-        res,
-        400,
-        "toAddress, amount, and assetSymbol are required",
-      );
-      return true;
-    }
-
-    const config = loadElizaConfig();
-    const tradePermissionMode = _resolveTradePermissionMode(config);
-    const canExecuteLocally = _canUseLocalTradeExecution(
-      tradePermissionMode,
-      isAgentAutomationRequest(req),
-    );
-    const hasLocalKey = Boolean(process.env.EVM_PRIVATE_KEY?.trim());
-    const hasStewardSigner = isStewardConfigured();
-    const canSign = hasLocalKey || hasStewardSigner;
-    const addresses = getWalletAddresses();
-    const rpcReadiness = resolveWalletRpcReadiness(config);
-    const bscExecutionNetwork = resolveBscExecutionNetwork();
-
-    let toAddress: string;
-    try {
-      toAddress = ethers.getAddress(toAddressRaw);
-    } catch {
-      sendJsonErrorResponse(
-        res,
-        400,
-        "Invalid toAddress — must be a valid EVM address",
-      );
-      return true;
-    }
-
-    const isBnb = assetSymbol.toUpperCase() === "BNB";
-    let decimals = 18;
-    if (typeof body.tokenAddress === "string" && body.tokenAddress.trim()) {
-      const provider = new ethers.JsonRpcProvider(
-        resolvePrimaryBscRpcUrl({
-          rpcUrls: rpcReadiness.bscRpcUrls,
-          cloudManagedAccess: rpcReadiness.cloudManagedAccess,
-        }) ?? "https://bsc-dataseed1.binance.org/",
-      );
-      try {
-        const tokenContract = new ethers.Contract(
-          body.tokenAddress,
-          ["function decimals() view returns (uint8)"],
-          provider,
-        );
-        decimals = Number(await tokenContract.decimals());
-      } finally {
-        provider.destroy();
-      }
-    }
-
-    const unsignedTx = {
-      chainId: bscExecutionNetwork.chainId,
-      from: addresses.evmAddress ?? null,
-      to:
-        isBnb || typeof body.tokenAddress !== "string"
-          ? toAddress
-          : body.tokenAddress,
-      data: isBnb
-        ? "0x"
-        : new ethers.Interface([
-            "function transfer(address to, uint256 amount) returns (bool)",
-          ]).encodeFunctionData("transfer", [
-            toAddress,
-            ethers.parseUnits(amount, decimals),
-          ]),
-      valueWei: isBnb ? ethers.parseEther(amount).toString() : "0",
-      explorerUrl: bscExecutionNetwork.explorerBaseUrl,
-      assetSymbol,
-      amount,
-      tokenAddress:
-        typeof body.tokenAddress === "string" ? body.tokenAddress : undefined,
-    };
-
-    if (!canSign || !canExecuteLocally || body.confirm !== true) {
-      sendJsonResponse(res, 200, {
-        ok: true,
-        mode: resolveWalletExecutionMode(
-          canSign,
-          canExecuteLocally,
-          hasStewardSigner,
-        ),
-        executed: false,
-        requiresUserSignature: true,
-        toAddress,
-        amount,
-        assetSymbol,
-        tokenAddress: unsignedTx.tokenAddress,
-        unsignedTx,
-      });
-      return true;
-    }
-
-    const _rpcUrl = resolvePrimaryBscRpcUrl({
-      rpcUrls: rpcReadiness.bscRpcUrls,
-      cloudManagedAccess: rpcReadiness.cloudManagedAccess,
-    });
-
-    try {
-      let finalHash = "";
-      let finalNonce: number | null = null;
-      let finalGasLimit = "0";
-      let finalMode: "local-key" | "steward" = hasLocalKey
-        ? "local-key"
-        : "steward";
-
-      if (hasLocalKey && canExecuteLocally) {
-        if (!_rpcUrl) {
-          sendJsonErrorResponse(
-            res,
-            503,
-            "BSC RPC not configured for local execution",
-          );
-          return true;
-        }
-
-        const localExecution = await _sendLocalWalletTransaction(_rpcUrl, {
-          to: unsignedTx.to,
-          data: unsignedTx.data,
-          value: BigInt(unsignedTx.valueWei),
-          chainId: unsignedTx.chainId,
-        });
-        finalHash = localExecution.hash;
-        finalNonce = localExecution.nonce;
-        finalGasLimit = localExecution.gasLimit;
-      } else {
-        finalMode = "steward";
-        const executionResult = await signTransactionWithOptionalSteward({
-          evmAddress: addresses.evmAddress,
-          tx: {
-            to: unsignedTx.to,
-            data: unsignedTx.data,
-            value: unsignedTx.valueWei,
-            chainId: unsignedTx.chainId,
-            broadcast: true,
-          },
-        });
-
-        if (
-          executionResult.mode === "steward" &&
-          executionResult.pendingApproval
-        ) {
-          sendJsonResponse(res, 200, {
-            ok: true,
-            mode: "steward",
-            executed: false,
-            requiresUserSignature: false,
-            toAddress,
-            amount,
-            assetSymbol,
-            tokenAddress: unsignedTx.tokenAddress,
-            unsignedTx,
-            execution: {
-              status: "pending_approval",
-              policyResults: executionResult.policyResults,
-            },
-          });
-          return true;
-        }
-
-        finalHash = "txHash" in executionResult ? executionResult.txHash : "";
-      }
-
-      sendJsonResponse(res, 200, {
-        ok: true,
-        mode: finalMode,
-        executed: true,
-        requiresUserSignature: false,
-        toAddress,
-        amount,
-        assetSymbol,
-        tokenAddress: unsignedTx.tokenAddress,
-        unsignedTx,
-        execution: {
-          hash: finalHash,
-          nonce: finalNonce,
-          gasLimit: finalGasLimit,
-          valueWei: unsignedTx.valueWei,
-          explorerUrl: `${bscExecutionNetwork.explorerBaseUrl}/tx/${finalHash}`,
-          blockNumber: null,
-          status: "pending",
-        },
-      });
-    } catch (err) {
-      if (isStewardPolicyRejection(err)) {
-        sendJsonResponse(res, 403, {
-          ok: false,
-          mode: "steward",
-          executed: false,
-          requiresUserSignature: false,
-          error: err.message,
-          execution: {
-            status: "rejected",
-            policyResults: getStewardPolicyResults(err),
-          },
-        });
-        return true;
-      }
-
-      sendJsonErrorResponse(
-        res,
-        500,
-        `Transfer failed: ${err instanceof Error ? err.message : "unknown error"}`,
-      );
-    }
-    return true;
-  }
+  // Wallet trade / transfer routes — extracted to wallet-trade-compat-routes.ts
+  if (await handleWalletTradeCompatRoutes(req, res, state)) return true;
 
   // Plugin routes — extracted to plugins-compat-routes.ts
   if (await handlePluginsCompatRoutes(req, res, state)) return true;
